+++ santashop-admin/src/app/components/qr-modal/qr-modal.component.html
