+++ santashop-admin/src/app/features/admin/scanner/scanner.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { IonicModule } from '@ionic/angular';
 import { ZXingScannerModule } from '@zxing/ngx-scanner';
-import { SharedComponentsModule } from 'santashop-admin/src/app/shared/components/shared-components.module';
+import { SharedComponentsModule } from 'santashop-admin/src/components/shared-components.module';
 import { ScannerPageRoutingModule } from './scanner-routing.module';
 import { ScannerPage } from './scanner.page';
