+++ santashop-admin/src/app/features/admin/scanner/scanner.page.ts
@@ -3,8 +3,8 @@ import { ModalController } from '@ionic/angular';
 import { ZXingScannerComponent } from '@zxing/ngx-scanner';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { distinctUntilChanged, map, publishReplay, refCount, takeUntil } from 'rxjs/operators';
-import { QrModalComponent } from 'santashop-admin/src/app/shared/components/qr-modal/qr-modal.component';
-import { CheckInService } from 'santashop-admin/src/app/shared/services/check-in.service';
+import { QrModalComponent } from 'santashop-admin/src/app/components/qr-modal/qr-modal.component';
+import { CheckInService } from 'santashop-admin/src/app/services/check-in.service';
 
 @Component({
   selector: 'app-scanner',
