+++ santashop-admin/src/app/features/sign-in/sign-in.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
-import { SharedComponentsModule } from '../../shared/components/shared-components.module';
+import { SharedComponentsModule } from 'santashop-admin/src/components/shared-components.module';
 import { SignInPageRoutingModule } from './sign-in-routing.module';
 import { SignInPage } from './sign-in.page';
