+++ santashop-admin/src/app/forms/quick-registration.ts
