+++ santashop-admin/src/app/helpers/registration-helpers.ts
